requirements.txt
django
python-google-places
djangorestframework
googlemaps